+++ packages/amplify-e2e-tests/src/__tests__/hosting.test.ts
@@ -1,13 +1,15 @@
-import { initJSProjectWithProfile, deleteProject } from '../init';
 import {
-  addHosting,
-  removeHosting,
   amplifyPushWithUpdate,
   amplifyPublishWithoutUpdate,
   createReactTestProject,
   resetBuildCommand,
-} from '../categories/hosting';
-import { deleteProjectDir, getProjectMeta } from '../utils';
+} from 'amplify-e2e-core';
+
+import { initJSProjectWithProfile, deleteProject } from 'amplify-e2e-core';
+import { addHosting, removeHosting, amplifyPushWithoutCodegen } from 'amplify-e2e-core';
+import { deleteProjectDir, getProjectMeta } from 'amplify-e2e-core';
+import * as fs from 'fs-extra';
+import * as path from 'path';
 
 describe('amplify add hosting', () => {
   let projRoot: string;
@@ -21,7 +23,7 @@ describe('amplify add hosting', () => {
 
   afterAll(async () => {
     await removeHosting(projRoot);
-    await amplifyPushWithUpdate(projRoot);
+    await amplifyPushWithoutCodegen(projRoot);
     await deleteProject(projRoot, true);
     deleteProjectDir(projRoot);
   });
@@ -52,4 +54,14 @@ describe('amplify add hosting', () => {
     expect(error.message).toEqual('Process exited with non zero exit code 1');
     resetBuildCommand(projRoot, currentBuildCommand);
   });
-});
+
+  it('add hosting', async () => {
+    await initJSProjectWithProfile(projRoot, {});
+    await addHosting(projRoot);
+    await amplifyPushWithoutCodegen(projRoot);
+    expect(fs.existsSync(path.join(projRoot, 'amplify', 'backend', 'hosting', 'S3AndCloudFront'))).toBe(true);
+    const projectMeta = getProjectMeta(projRoot);
+    expect(projectMeta.hosting).toBeDefined();
+    expect(projectMeta.hosting.S3AndCloudFront).toBeDefined();
+  });
+})