+++ packages/amplify-e2e-tests/src/init/index.ts
@@ -1,5 +1,0 @@
-export * from './initProjectHelper';
-export * from './amplifyPush';
-export * from './deleteProject';
-export * from './pull-headless';
-export { getProjectMeta } from '../utils';